+++ examples/complete/variables.tf
@@ -1,10 +1,10 @@
 variable "name" {
-  description = " Name of the cluster (up to 255 letters, numbers, hyphens, and underscores)"
+  description = "Name of the cluster (up to 255 letters, numbers, hyphens, and underscores)"
   type        = string
   default     = "complete-ecs-cluster-example"
 }
 variable "supporting_resources_name" {
-  description = " Name of the supporting resources name tag"
+  description = "Name of the supporting resources name tag"
   type        = string
   default     = "terraform-aws-ecs-cluster"
 }
@@ -22,9 +22,9 @@ variable "create_ec2_instance" {
 }
 
 variable "instance_type" {
-  description = "(Optional) The type of the instance."
+  description = "The type of the instance."
   type        = string
-  default     = "t2.micro"
+  default     = "t3.medium"
 }
 
 variable "associate_public_ip_address" {
@@ -45,31 +45,6 @@ variable "logging" {
   default     = "OVERRIDE"
 }
 
-variable "block_device_mappings" {
-  description = "(Optional) Specify volumes to attach to the instance besides the volumes specified by the AMI. "
-  type        = list(any)
-  default = [
-    {
-      device_name = "/dev/xvda"
-      ebs = {
-        delete_on_termination = true
-        volume_size           = 20
-        volume_type           = "gp2"
-        encrypted             = true
-      }
-    },
-    {
-      device_name = "/dev/xvdcz"
-      ebs = {
-        delete_on_termination = true
-        volume_size           = 22
-        volume_type           = "gp2"
-        encrypted             = true
-      }
-    }
-  ]
-}
-
 variable "key_description" {
   description = "The description of the key as viewed in AWS console."
   type        = string
@@ -77,7 +52,7 @@ variable "key_description" {
 }
 
 variable "tags" {
-  description = " A map of tags to assign to the object. If configured with a provider default_tags configuration block present, tags with matching keys will overwrite those defined at the provider-level."
+  description = "A map of tags to assign to the object. If configured with a provider default_tags configuration block present, tags with matching keys will overwrite those defined at the provider-level."
   type        = map(string)
   default = {
     Environment        = "example"
@@ -94,5 +69,87 @@ variable "tags" {
 variable "create_kms_key" {
   description = "Whether or not to create a kms key with this module"
   type        = bool
+  default     = true
+}
+
+variable "monitoring_enabled" {
+  description = "The monitoring option for the instance."
+  type        = bool
+  default     = true
+}
+
+# ECS Service
+
+variable "image" {
+  type        = string
+  description = "Name of image to pull from dockerhub"
+  default     = "boldlink/flaskapp:latest"
+}
+
+variable "cpu" {
+  type        = number
+  description = "The number of cpu units to allocate"
+  default     = 10
+}
+
+variable "memory" {
+  type        = number
+  description = "The size of memory to allocate in MiBs"
+  default     = 512
+}
+
+variable "essential" {
+  type        = bool
+  description = "Whether this container is essential"
+  default     = true
+}
+
+variable "containerport" {
+  type        = number
+  description = "Specify container port"
+  default     = 5000
+}
+
+variable "hostport" {
+  type        = number
+  description = "Specify host port"
+  default     = 5000
+}
+
+variable "network_mode" {
+  type        = string
+  description = "Docker networking mode to use for the containers in the task. Valid values are none, bridge, awsvpc, and host."
+  default     = "bridge"
+}
+
+variable "service_ingress_rules" {
+  description = "Ingress rules to add to the service security group."
+  type        = list(any)
+  default = [
+    {
+      from_port   = 5000
+      to_port     = 5000
+      protocol    = "tcp"
+      description = "Allow traffic on port 5000. The app is configured to use this port"
+      cidr_blocks = ["0.0.0.0/0"]
+    }
+  ]
+}
+
+variable "launch_type" {
+  default     = "EC2"
+  description = "Launch type on which to run your service. The valid values are EC2, FARGATE, and EXTERNAL. Defaults to EC2."
+  type        = string
+}
+
+variable "requires_compatibilities" {
+  description = "Set of launch types required by the task. The valid values are EC2 and FARGATE."
+  type        = list(string)
+  default     = []
+}
+
+variable "install_ssm_agent" {
+  type        = bool
+  description = "Whether to install ssm agent"
   default     = true
 }